TST/CLN: move some tests
